# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List products",
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create product",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/categories": {
      "get": {
        "operationId": "ProductsController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List product categories",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/barcode/{barcode}": {
      "get": {
        "operationId": "ProductsController_findByBarcode",
        "parameters": [
          {
            "name": "barcode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get product by barcode",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get product by ID",
        "tags": [
          "Products"
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update product",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete product",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/inventory/product/{productId}": {
      "get": {
        "operationId": "InventoryController_findByProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get inventory by product ID",
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/orders": {
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List orders",
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/bills/all": {
      "get": {
        "operationId": "OrdersController_findAllBills",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List bills",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/bills": {
      "post": {
        "operationId": "OrdersController_createBill",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create bill for an order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/bills/{billNo}": {
      "get": {
        "operationId": "OrdersController_findOneBill",
        "parameters": [
          {
            "name": "billNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get bill by bill number",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/payments/all": {
      "get": {
        "operationId": "OrdersController_findAllPayments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List payments",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/payments": {
      "post": {
        "operationId": "OrdersController_createPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Record payment for a bill",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/payments/{billNo}": {
      "get": {
        "operationId": "OrdersController_findOnePayment",
        "parameters": [
          {
            "name": "billNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get payment by bill number",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get order by ID",
        "tags": [
          "Orders"
        ]
      },
      "put": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update order status or payment details",
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete order",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/returns": {
      "get": {
        "operationId": "ReturnsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "List returns",
        "tags": [
          "Returns"
        ]
      },
      "post": {
        "operationId": "ReturnsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create return request",
        "tags": [
          "Returns"
        ]
      }
    },
    "/api/returns/{id}": {
      "get": {
        "operationId": "ReturnsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get return by ID",
        "tags": [
          "Returns"
        ]
      },
      "put": {
        "operationId": "ReturnsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReturnDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update return status/details",
        "tags": [
          "Returns"
        ]
      },
      "delete": {
        "operationId": "ReturnsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete return request",
        "tags": [
          "Returns"
        ]
      }
    },
...
  },
  "info": {
    "title": "BillBhai API",
    "description": "The BillBhai Retail Order Processing System API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "string",
            "example": "SUP-001",
            "description": "Supplier ID"
          },
          "name": {
            "type": "string",
            "example": "Basmati Rice",
            "description": "Product name"
          },
          "category": {
            "type": "string",
            "example": "Groceries",
            "description": "Product category"
          },
          "barcode": {
            "type": "string",
            "example": "BAR001",
            "description": "Barcode"
          },
          "price": {
            "type": "number",
            "example": 380,
            "description": "Product price",
            "minimum": 0
          },
          "size": {
            "type": "string",
            "example": "5kg",
            "description": "Product size"
          },
          "description": {
            "type": "string",
            "example": "Premium basmati rice",
            "description": "Product description"
          }
        },
        "required": [
          "supplierId",
          "name",
          "category",
          "price"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "string",
            "example": "SUP-001",
            "description": "Supplier ID"
          },
          "name": {
            "type": "string",
            "example": "Basmati Rice",
            "description": "Product name"
          },
          "category": {
            "type": "string",
            "example": "Groceries",
            "description": "Product category"
          },
          "barcode": {
            "type": "string",
            "example": "BAR001",
            "description": "Barcode"
          },
          "price": {
            "type": "number",
            "example": 400,
            "description": "Product price",
            "minimum": 0
          },
          "size": {
            "type": "string",
            "example": "5kg",
            "description": "Product size"
          },
          "description": {
            "type": "string",
            "example": "Premium basmati rice",
            "description": "Product description"
          }
        }
      },
...
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "P001",
            "description": "Product ID"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity ordered",
            "minimum": 1
          },
          "itemPrice": {
            "type": "number",
            "example": 380,
            "description": "Price per item",
            "minimum": 0
          }
        },
        "required": [
          "productId",
          "quantity",
          "itemPrice"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "example": "CUS-001",
            "description": "Customer ID"
          },
          "staffId": {
            "type": "string",
            "example": "USR-002",
            "description": "Staff ID"
          },
          "companyId": {
            "type": "string",
            "example": "BIZ-101",
            "description": "Company ID"
          },
          "orderType": {
            "type": "string",
            "enum": [
              "pickup",
              "delivery"
            ],
            "description": "Order type"
          },
          "checkoutMode": {
            "type": "string",
            "enum": [
              "takeaway_now",
              "prepaid_delivery",
              "cod_delivery"
            ],
            "description": "Checkout mode"
          },
          "discountAmount": {
            "type": "number",
            "example": 100,
            "description": "Discount amount",
            "minimum": 0
          },
          "paymentMethod": {
            "type": "string",
            "example": "UPI",
            "description": "Payment method"
          },
          "items": {
            "description": "Order items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          }
        },
        "required": [
          "customerId",
          "staffId",
          "companyId",
          "orderType",
          "checkoutMode",
          "items"
        ]
      },
      "CreateBillDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "example": "ORD-4821",
            "description": "Order ID"
          },
          "taxAmount": {
            "type": "number",
            "example": 50,
            "description": "Tax amount",
            "minimum": 0
          },
          "discountAmount": {
            "type": "number",
            "example": 0,
            "description": "Discount amount",
            "minimum": 0
          }
        },
        "required": [
          "orderId"
        ]
      },
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "billNo": {
            "type": "string",
            "example": "BILL-001",
            "description": "Bill number"
          },
          "paymentMethod": {
            "type": "string",
            "example": "Card",
            "description": "Payment method"
          },
          "amountPaid": {
            "type": "number",
            "example": 1250,
            "description": "Amount paid",
            "minimum": 0
          }
        },
        "required": [
          "billNo",
          "paymentMethod",
          "amountPaid"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "Pending",
              "Processing",
              "Delivered",
              "Cancelled"
            ],
            "description": "Order status"
          },
          "paymentMethod": {
            "type": "string",
            "example": "Card",
            "description": "Payment method"
          }
        }
      },
...
      "CreateReturnDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "example": "ORD-4821",
            "description": "Order ID"
          },
          "staffId": {
            "type": "string",
            "example": "USR-002",
            "description": "Staff ID"
          },
          "reason": {
            "type": "string",
            "example": "Damaged product",
            "description": "Return reason"
          },
          "refundAmount": {
            "type": "number",
            "example": 100,
            "description": "Refund amount",
            "minimum": 0
          },
          "returnType": {
            "type": "string",
            "example": "refund",
            "description": "Return type"
          },
          "product": {
            "type": "string",
            "example": "P001",
            "description": "Product ID"
          },
          "qty": {
            "type": "number",
            "example": 1,
            "description": "Quantity returned",
            "minimum": 1
          },
          "requestedBy": {
            "type": "string",
            "example": "Arjun Mehta",
            "description": "Name of person requesting return"
          }
        },
        "required": [
          "orderId",
          "staffId",
          "reason",
          "refundAmount",
          "returnType",
          "product",
          "qty",
          "requestedBy"
        ]
      },
      "UpdateReturnDto": {
        "type": "object",
        "properties": {
          "refundAmount": {
            "type": "number",
            "example": 100,
            "description": "Refund amount",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "Pending",
              "Approved",
              "Refunded",
              "Rejected",
              "Under Review"
            ],
            "description": "Return status"
          },
          "reason": {
            "type": "string",
            "example": "Damaged product",
            "description": "Return reason"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "x-role": []
    }
  ]
}